This adds a text dump function (press shift-T) in the OGL backend, saving a big text file to /tmp/dumpframe####.txt with the numbers counting up.
The data for each vertex is dumped as well as the table of index values
for each glDrawElements() call. Unfortunately there isn't enough information
to recreate the rendered output. What's missing are:
1) The vertex and pixel shaders (the geometry shader also perhaps)
2) The data passed as uniforms to the shader programs, which includes the
   matrices for transforming the position coordinates (and texture coordinates
   I suppose). Lighting sources are also not included.
3) The textures used.

This approach is unworkable. I'm switching to a binary format where I just
duplicate the structures present in memory. I'll also include the uniform
data and the text of the shaders, and I'll make it so the draw primitive
element tells which shader program to use.

I'll worry about textures later. I'm hoping after doing the above I'll at
least have everything recognizable and in the right place...